backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
pydantic==2.10.3
pydantic-settings==2.6.1
firebase-admin==6.5.0
google-generativeai==0.8.3
PyMuPDF==1.24.14
chromadb==0.5.23